Upgrading react navigation to version 6.x.x
