+++ package.json
@@ -11,9 +11,9 @@
     "@expo/metro-config": "^0.1.60",
     "@react-native-community/checkbox": "^0.5.8",
     "@react-native-community/masked-view": "0.1.10",
-    "@react-navigation/bottom-tabs": "^5.11.8",
-    "@react-navigation/native": "^5.9.3",
-    "@react-navigation/stack": "^5.14.4",
+    "@react-navigation/bottom-tabs": "^6.0.7",
+    "@react-navigation/native": "^6.0.4",
+    "@react-navigation/stack": "^6.0.9",
     "axios": "^0.21.4",
     "expo": "~40.0.0",
     "expo-linear-gradient": "^8.4.0",
