+++ src/Routes/MainBottom/RoutesName.ts
@@ -1,0 +1,6 @@
+export const ROUTES_NAME = {
+    HOME: "ActivitySelection",
+    SETTINGS: "Settings",
+    USER: "User",
+    ACTIVITY: "Activity",
+  };
